requirements.txt
typer
rich
pandas
python-magic
python-dotenv
loguru
gspread
google-auth
